Refactored feeder protocol class to IndexCommandEncoder
